refactoring and unittest
